maybe fix valkyrieobject load issue
